+++ src/EventSourcing/Logitar.EventSourcing.EntityFrameworkCore.SqlServer/DependencyInjectionExtensions.cs
@@ -19,7 +19,8 @@ public static class DependencyInjectionExtensions
   {
     return services
       .AddDbContext<EventContext>(options => options.UseSqlServer(connectionString,
-        options => options.MigrationsAssembly("Logitar.EventSourcing.EntityFrameworkCore.SqlServer")))
-      .AddScoped<IAggregateRepository, AggregateRepository>();
+        options => options.MigrationsAssembly("Logitar.EventSourcing.EntityFrameworkCore.SqlServer")
+      ))
+      .AddLogitarEventSourcingWithEntityFrameworkCoreRelational();
   }
 }