+++ src/EventSourcing/Logitar.EventSourcing.InMemory/AggregateRepository.cs
@@ -6,7 +6,8 @@ public class AggregateRepository : Infrastructure.AggregateRepository
 {
   private readonly List<EventEntity> _events = new();
 
-  public AggregateRepository(IEventBus eventBus) : base(eventBus)
+  public AggregateRepository(IEventBus eventBus, IEventSerializer eventSerializer)
+    : base(eventBus, eventSerializer)
   {
   }
 
@@ -18,7 +19,7 @@ public class AggregateRepository : Infrastructure.AggregateRepository
     IEnumerable<DomainEvent> changes = _events.Where(e => e.AggregateType == aggregateType && e.AggregateId == aggregateId
         && (!version.HasValue || version.Value <= e.Version))
       .OrderBy(e => e.Version)
-      .Select(EventSerializer.Instance.Deserialize);
+      .Select(EventSerializer.Deserialize);
 
     return Task.FromResult(changes);
   }
@@ -29,7 +30,7 @@ public class AggregateRepository : Infrastructure.AggregateRepository
 
     IEnumerable<DomainEvent> changes = _events.Where(e => e.AggregateType == aggregateType)
       .OrderBy(e => e.Version)
-      .Select(EventSerializer.Instance.Deserialize);
+      .Select(EventSerializer.Deserialize);
 
     return Task.FromResult(changes);
   }
@@ -41,14 +42,14 @@ public class AggregateRepository : Infrastructure.AggregateRepository
 
     IEnumerable<DomainEvent> changes = _events.Where(e => e.AggregateType == aggregateType && aggregateIds.Contains(e.AggregateId))
       .OrderBy(e => e.Version)
-      .Select(EventSerializer.Instance.Deserialize);
+      .Select(EventSerializer.Deserialize);
 
     return Task.FromResult(changes);
   }
 
   protected override Task SaveChangesAsync(IEnumerable<AggregateRoot> aggregates, CancellationToken cancellationToken)
   {
-    IEnumerable<EventEntity> events = aggregates.SelectMany(EventEntity.FromChanges);
+    IEnumerable<EventEntity> events = aggregates.SelectMany(aggregate => EventEntity.FromChanges(aggregate, EventSerializer));
     _events.AddRange(events);
 
     return Task.CompletedTask;